Added lambertian material and more SoA code

+++ src/Private/Camera.cpp
@@ -64,7 +64,7 @@ Color Camera::PerformPathTrace(const Ray& R, HittableList& World) const
 			Vector3D UnitDirection = CurrentRay.Direction().Normalize();
 			float t = 0.5f * (UnitDirection.X + 1.f);//We are working with a unit vector with X in [-1,1] so we have to map X from [-1,1] to [0,1] first
 			PixelColor += ((1.f - t) * Color(0.9f, 0.9f, 0.9f) + t * Color(0.5f, 0.7f, 1.f));
-			return std::pow(0.5f, i) * PixelColor;
+			return std::powf(0.5f, i) * PixelColor;
 		}
 	}
